Refactor: Move check_require_cache code to helpers folder
The root of the `datadog-instrumentations` package is reserved for
module instrumentation code. The `check_require_cache.js` file is a
helper file and should be in a sub-directory.

This commit also consolidates the `helpers` and `utils` folders as they
served the same purpose.
