+++ .github/workflows/instrumentations.yml
@@ -45,14 +45,6 @@ jobs:
   # instrumentations directory, so they need to be run somewhere. This seems to
   # be a reasonable place to run them for now.
 
-  check_require_cache:
-    runs-on: ubuntu-latest
-    env:
-      PLUGINS: check_require_cache
-    steps:
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-      - uses: ./.github/actions/plugins/test
-
   express-session:
     runs-on: ubuntu-latest
     env:
