+++ packages/datadog-instrumentations/src/helpers/extract-package-and-module-path.js
