+++ packages/datadog-instrumentations/test/helpers/check-require-cache.spec.js
@@ -2,7 +2,7 @@
 
 const { exec } = require('node:child_process')
 
-describe('check_require_cache', () => {
+describe('check-require-cache', () => {
   const opts = {
     cwd: __dirname,
     env: {
@@ -11,7 +11,7 @@ describe('check_require_cache', () => {
   }
 
   it('should be no warnings when tracer is loaded first', (done) => {
-    exec(`${process.execPath} ./check_require_cache/good-order.js`, opts, (error, stdout, stderr) => {
+    exec(`${process.execPath} ./check-require-cache/good-order.js`, opts, (error, stdout, stderr) => {
       expect(error).to.be.null
       expect(stderr).to.not.include("Package 'express' was loaded")
       done()
@@ -21,7 +21,7 @@ describe('check_require_cache', () => {
   // stderr is empty on Windows
   if (process.platform !== 'windows') {
     it('should find warnings when tracer loaded late', (done) => {
-      exec(`${process.execPath} ./check_require_cache/bad-order.js`, opts, (error, stdout, stderr) => {
+      exec(`${process.execPath} ./check-require-cache/bad-order.js`, opts, (error, stdout, stderr) => {
         expect(error).to.be.null
         expect(stderr).to.include("Package 'express' was loaded")
         done()
