+++ packages/datadog-instrumentations/test/helpers/check-require-cache/bad-order.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 require('express') // package required before tracer
-const tracer = require('../../../../')
+const tracer = require('../../../../../')
 tracer.init()
 
 process.exit()
