+++ packages/datadog-instrumentations/test/helpers/check-require-cache/good-order.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const tracer = require('../../../../')
+const tracer = require('../../../../../')
 require('express') // package required after tracer
 tracer.init()
